requirements.txt
fastapi~=0.115.12
PyJWT~=2.10.1
passlib~=1.7.4
pydantic~=2.11.4
uvicorn~=0.34.2
python-dotenv~=1.1.0